refactor: unify ad class name and seperate service
