+++ src/main/java/com/myce/advertisement/controller/UserAdController.java
@@ -1,7 +1,7 @@
 package com.myce.advertisement.controller;
 
-import com.myce.advertisement.dto.AdvertisementRegistrationRequest;
-import com.myce.advertisement.service.UserAdvertisementService;
+import com.myce.advertisement.dto.AdRegistrationRequest;
+import com.myce.advertisement.service.UserAdService;
 import com.myce.auth.dto.CustomUserDetails;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
@@ -16,15 +16,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/api/ads")
 @RequiredArgsConstructor
-public class UserAdvertisementController {
-  private final UserAdvertisementService userAdvertisementService;
+public class UserAdController {
+  private final UserAdService userAdService;
 
   // 광고 등록
   @PostMapping
   public ResponseEntity<Long> saveAdvertisement(@AuthenticationPrincipal CustomUserDetails customUserDetails,
-      @RequestBody @Valid AdvertisementRegistrationRequest advertisementRegistrationRequest){
+      @RequestBody @Valid AdRegistrationRequest adRegistrationRequest){
     Long memberId = customUserDetails.getMemberId();
-    userAdvertisementService.saveAdvertisement(memberId, advertisementRegistrationRequest);
+    userAdService.saveAdvertisement(memberId, adRegistrationRequest);
     return ResponseEntity.status(HttpStatus.CREATED).build();
   }
 }
