+++ src/main/java/com/myce/advertisement/repository/AdRepository.java
@@ -12,7 +12,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface AdvertisementRepository extends JpaRepository<Advertisement, Long> {
+public interface AdRepository extends JpaRepository<Advertisement, Long> {
 
     Page<Advertisement> findByStatusIn(List<AdvertisementStatus> statuses, Pageable pageable);
