+++ src/main/java/com/myce/advertisement/service/PlatformAdDetailService.java
@@ -1,0 +1,8 @@
+package com.myce.advertisement.service;
+
+import com.myce.advertisement.dto.*;
+
+public interface PlatformAdDetailService {
+
+    AdDetailResponse getDetail(Long bannerId);
+}
