+++ src/main/java/com/myce/advertisement/service/PlatformCurrentAdService.java
@@ -1,0 +1,10 @@
+package com.myce.advertisement.service;
+
+import com.myce.advertisement.dto.AdCancelInfoCheck;
+import com.myce.advertisement.dto.AdCancelInfoRequest;
+
+public interface PlatformCurrentAdService {
+    void cancelBanner(Long bannerId, AdCancelInfoRequest request);
+
+    AdCancelInfoCheck generateCancelCheck(Long bannerId);
+}
