+++ src/main/java/com/myce/advertisement/service/UserAdService.java
@@ -1,0 +1,8 @@
+package com.myce.advertisement.service;
+
+
+import com.myce.advertisement.dto.AdRegistrationRequest;
+
+public interface UserAdService {
+  void saveAdvertisement(Long memberId, AdRegistrationRequest request);
+}
