--- src/main/java/com/myce/advertisement/service/UserAdvertisementService.java
@@ -1,8 +1,0 @@
-package com.myce.advertisement.service;
-
-
-import com.myce.advertisement.dto.AdvertisementRegistrationRequest;
-
-public interface UserAdvertisementService {
-  void saveAdvertisement(Long memberId, AdvertisementRegistrationRequest request);
-}
