+++ src/main/java/com/myce/advertisement/service/mapper/AdApplyMapper.java
@@ -1,0 +1,14 @@
+package com.myce.advertisement.service.mapper;
+
+import com.myce.advertisement.dto.AdPaymentHistoryResponse;
+import com.myce.advertisement.dto.AdPaymentInfoCheck;
+import com.myce.advertisement.dto.AdRejectInfoResponse;
+import com.myce.advertisement.entity.Advertisement;
+import com.myce.common.entity.RejectInfo;
+import com.myce.payment.entity.AdPaymentInfo;
+
+import java.util.HashMap;
+
+public class AdApplyMapper {
+
+}
