+++ src/main/java/com/myce/advertisement/service/mapper/AdRegistrationMapper.java
@@ -1,13 +1,13 @@
 package com.myce.advertisement.service.mapper;
 
-import com.myce.advertisement.dto.AdvertisementRegistrationRequest;
+import com.myce.advertisement.dto.AdRegistrationRequest;
 import com.myce.advertisement.entity.AdPosition;
 import com.myce.advertisement.entity.Advertisement;
 import com.myce.advertisement.entity.type.AdvertisementStatus;
 import com.myce.member.entity.Member;
 
-public class AdvertisementRegistrationMapper {
-  public static Advertisement toEntity(AdvertisementRegistrationRequest request, Member member, AdPosition adPosition, int totalDays) {
+public class AdRegistrationMapper {
+  public static Advertisement toEntity(AdRegistrationRequest request, Member member, AdPosition adPosition, int totalDays) {
     return Advertisement.builder()
         .member(member)
         .title(request.getTitle())
