+++ src/main/java/com/myce/member/entity/Member.java
@@ -100,6 +100,7 @@ public class Member {
   
     public void resetPassword(String newPassword) {
         this.password = newPassword;
+    }
     
     public void updateInfo(String phone, String email) {
         if (phone != null && !phone.trim().isEmpty()) {
