+++ src/main/java/com/myce/member/service/impl/MemberAdServiceImpl.java
@@ -1,7 +1,7 @@
 package com.myce.member.service.impl;
 
 import com.myce.advertisement.entity.Advertisement;
-import com.myce.advertisement.repository.AdvertisementRepository;
+import com.myce.advertisement.repository.AdRepository;
 import com.myce.common.entity.BusinessProfile;
 import com.myce.common.entity.type.TargetType;
 import com.myce.common.exception.CustomErrorCode;
@@ -23,7 +23,7 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class MemberAdServiceImpl implements MemberAdService {
 
-    private final AdvertisementRepository advertisementRepository;
+    private final AdRepository adRepository;
     private final MemberAdvertisementMapper memberAdvertisementMapper;
     private final AdvertisementDetailMapper advertisementDetailMapper;
     private final AdvertisementPaymentDetailMapper advertisementPaymentDetailMapper;
@@ -33,16 +33,16 @@ public class MemberAdServiceImpl implements MemberAdService {
 
     @Override
     public List<MemberAdvertisementResponse> getMemberAdvertisements(Long memberId) {
-        List<Advertisement> advertisements = advertisementRepository.findByMemberIdWithAdPosition(memberId);
+        List<Advertisement> advertisements = adRepository.findByMemberIdWithAdPosition(memberId);
         return memberAdvertisementMapper.toResponseDtoList(advertisements);
     }
 
     @Override
     public AdvertisementDetailResponse getAdvertisementDetail(Long memberId, Long advertisementId) {
 
-        Advertisement advertisement = advertisementRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
+        Advertisement advertisement = adRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
                         memberId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.ADVERTISEMENT_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(CustomErrorCode.AD_NOT_FOUND));
 
         BusinessProfile businessProfile = businessProfileRepository.findByTargetIdAndTargetType(advertisementId,
                         TargetType.ADVERTISEMENT)
@@ -54,9 +54,9 @@ public class MemberAdServiceImpl implements MemberAdService {
     @Override
     @Transactional
     public void cancelAdvertisement(Long memberId, Long advertisementId) {
-        Advertisement advertisement = advertisementRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
+        Advertisement advertisement = adRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
                         memberId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.ADVERTISEMENT_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(CustomErrorCode.AD_NOT_FOUND));
 
         advertisement.cancel();
     }
@@ -64,9 +64,9 @@ public class MemberAdServiceImpl implements MemberAdService {
     @Override
     public AdvertisementPaymentDetailResponse getAdvertisementPaymentDetail(Long memberId, Long advertisementId) {
         // 광고 정보 조회
-        Advertisement advertisement = advertisementRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
+        Advertisement advertisement = adRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
                         memberId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.ADVERTISEMENT_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(CustomErrorCode.AD_NOT_FOUND));
 
         // 사업자 정보 조회
         BusinessProfile businessProfile = businessProfileRepository.findByTargetIdAndTargetType(advertisementId,
@@ -83,9 +83,9 @@ public class MemberAdServiceImpl implements MemberAdService {
     @Override
     public AdvertisementRefundReceiptResponse getAdvertisementRefundReceipt(Long memberId, Long advertisementId) {
         // 광고 정보 조회
-        Advertisement advertisement = advertisementRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
+        Advertisement advertisement = adRepository.findByIdAndMemberIdWithAdPosition(advertisementId,
                         memberId)
-                .orElseThrow(() -> new CustomException(CustomErrorCode.ADVERTISEMENT_NOT_FOUND));
+                .orElseThrow(() -> new CustomException(CustomErrorCode.AD_NOT_FOUND));
 
         // 게시 중인 광고만 환불 가능
         if (advertisement.getStatus() != com.myce.advertisement.entity.type.AdvertisementStatus.PUBLISHED) {
